+++ src/world/level.cpp
@@ -54,4 +54,16 @@ void Level::load(const Vector3& pos, const Quaternion& rot)
 	}
 }
 
+UnitId Level::unit_by_name(StringId32 name)
+{
+	const StringId32* unit_names = level_resource::unit_names(_resource);
+	for (u32 i = 0; i < _resource->num_units; ++i)
+	{
+		if (unit_names[i] == name)
+			return _unit_lookup[i];
+	}
+
+	return UNIT_INVALID;
+}
+
 } // namespace crown